Title: 0에 가장 가까운 합, Time: 455ms, Memory: 22MB, Status: Wrong Answer - Codetree

+++ 251027/0에 가장 가까운 합/sum-closest-to-zero.java
@@ -39,8 +39,8 @@ class Solver {
         int i = 0;
         int j = numbers.size() - 1;
         Collections.sort(numbers);
-        int sum = Math.abs(numbers.get(i) + numbers.get(j));
-        int answer = sum;
+        long sum = Math.abs(numbers.get(i) + numbers.get(j));
+        long answer = sum;
         loop:
         for (i = 0; i < numbers.size() - 1; i++) {
             while (true) {
